JavaBasics/src/main/java/com/sparta/mg/Program.java
package com.sparta.mg;
//Read, Maintain, Test
public class Program {
    public static void main(String[] args) { //entry point (methods)
        int timeOfDay = 12; //type, name, value.
        System.out.println(getGreeting(timeOfDay));
    }

    public static String getGreeting(int timeOfDay) {
        String greeting;
        if (timeOfDay >= 5 && timeOfDay <= 12) {
            greeting = "Good morning!";
        } else if (timeOfDay >= 12 && timeOfDay <= 18) {
            greeting = "Good afternoon!";
        } else {
            greeting = "Good evening!";
        }
        return greeting;
    }
}